add keeper cli and common packages
