+++ packages/cli/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@marginly/deploy": "0.1.0",
+    "@marginly/common": "0.1.0",
     "@ethersproject/hardware-wallets": "^5.6.0",
     "@marginly/contracts": "0.1.0",
     "@uniswap/v3-core": "^1.0.1",
