+++ packages/common/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@marginly/common",
+  "version": "0.1.0",
+  "main": "dist/index.js",
+  "author": "EqLab",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc --build"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.31.0",
+    "@typescript-eslint/parser": "^4.31.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "prettier": "^2.4.0",
+    "typescript": "^4.4.2"
+  },
+  "dependencies": {
+    "ethers": "^5.7.2",
+    "fs-extra": "^11.1.0",
+    "commander": "^8.1.0",
+    "ganache": "^7.7.3",
+    "prompts": "^2.4.2",
+    "read": "^2.0.0"
+  }
+}
