+++ packages/common/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+      "outDir": "dist",
+      "declaration": true,
+      "composite": true,
+      "rootDir": "src",
+      "resolveJsonModule": true
+    },
+    "include": ["src/**/*", "src/**/*.json"],
+  }
+
