+++ packages/keeper/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@marginly/keeper",
+  "version": "0.1.0",
+  "description": "Marginly keeper service",
+  "main": "index.js",
+  "scripts": {
+    "build": "tsc --build",
+    "dev": "ts-node src/index.ts",
+    "start": "node dist/index.js",
+    "lint": "eslint src --ext .ts",
+    "prettify": "prettier --write ."
+  },
+  "keywords": [],
+  "author": "EqLab",
+  "license": "MIT",
+  "dependencies": {
+    "@marginly/common": "0.1.0"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.31.0",
+    "@typescript-eslint/parser": "^4.31.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "prettier": "^2.4.0",
+    "ts-node": "^10.9.1",
+    "@types/node": "^18.15.11",
+    "typescript": "^5.0.4"
+  }
+}
