+++ packages/keeper/src/data/ethereum/config.json
@@ -1,0 +1,19 @@
+{
+  "systemContextDefaults": {
+    "ethNodeUri": "https://ethereum-mainnet-rpc.allthatnode.com"
+  },
+  "aavePoolAdressesProviderAddress": "0x2f39d218133AFaB8F2B819B1066c7E434Ad94E9e",
+  "uniswapRouterContractAddress": "0xE592427A0AEce92De3Edee1F18E0157C05861564",
+  "marginlyPools": [
+    {
+      "address": "0x...",
+      "minProfitQuote": "",
+      "minProfitBase": ""
+    },
+    {
+      "address": "0x...",
+      "minProfitQuote": "",
+      "minProfitBase": ""
+    }
+  ]
+}
